pyproject.toml
[project]
name = "work-rpa"
version = "0.1.0"
description = "FAX業務自動化RPAシステム - AI-OCRエージェントチーム"
requires-python = ">=3.10"
dependencies = [
  "streamlit>=1.30.0",
  "streamlit-drawable-canvas>=0.9.3",
  "opencv-python>=4.8.0",
  "Pillow>=10.0.0",
  "pytesseract>=0.3.10",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "python-dotenv>=1.0.0",
  "numpy>=1.24.0",
]

[project.optional-dependencies]
azure = [
  "azure-ai-documentintelligence>=1.0.0",
]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
